+++ pages/api/prospects.js
@@ -1,11 +1,10 @@
-import { register } from '../../api/prospects'
+import { register } from '../../services/prospects'
 
 async function list () {
   return []
 }
 
 export default async function handler (req, res) {
-  
   try {
     switch (req.method) {
       case 'GET':
